+++ src/views/franchise/OrderDetailView.vue
@@ -9,6 +9,7 @@
           class="mb-1"
         />
         <div class="top-buttons">
+          <Button label="주문요청서 출력" variant="outlined" size="small" @click="clickPrintOrder" />
           <Button label="교환" variant="outlined" size="small" :disabled="!isCompleted" @click="clickExchange" />
           <Button label="반품" variant="outlined" size="small" :disabled="!isCompleted" @click="clickReturn" />
           <Button label="목록으로" size="small" severity="secondary" variant="outlined" @click="clickGoToList" />
@@ -26,44 +27,10 @@
       <div class="body-area">
         <h1>주문상세</h1>
 
-        <AppTableStyled full-width>
-          <tbody>
-            <tr>
-              <th>주문번호</th>
-              <td>{{ orderDetail.orderCode }}</td>
-              <th>주문금액</th>
-              <td>{{ orderDetail.sumPrice.toLocaleString() }}</td>
-              <th>주문일자</th>
-              <td>{{ orderDetail.createdAt }}</td>
-              <th>완료일자</th>
-              <td>{{ orderDetail.doneDate }}</td>
-            </tr>
-            <tr>
-              <th colspan="8">주문품목</th>
-            </tr>
-            <tr>
-              <th>품목코드</th>
-              <th colspan="3">품목명</th>
-              <th>수량</th>
-              <th>단가</th>
-              <th>주문금액</th>
-              <th>부가세</th>
-            </tr>
-            <tr v-for="item in orderDetail.orderItemList" :key="item.itemCode">
-              <td class="align-center">{{ item.itemCode }}</td>
-              <td colspan="3">{{ item.name }}</td>
-              <td class="align-right">{{ item.quantity.toLocaleString() }}</td>
-              <td class="align-right">{{ (item.partSum / item.quantity).toLocaleString() }}</td>
-              <td class="align-right">{{ item.partSum.toLocaleString() }}</td>
-              <td class="align-right">{{ (item.partSum * 0.1).toLocaleString() }}</td>
-            </tr>
-            <tr>
-              <th>총 주문금액</th>
-              <td class="align-right" colspan="7">{{ totalPrice.toLocaleString() }}</td>
-            </tr>
-          </tbody>
-        </AppTableStyled>
+        <OrderDetailTable :order-detail="orderDetail" />
       </div>
+
+      <PrintOrderPdfPreviewModal v-model:show="showPrintPdf" :order-detail="orderDetail" />
     </template>
   </div>
 </template>
@@ -73,7 +40,8 @@ import { useToast } from 'primevue';
 import { computed, onMounted, ref } from 'vue';
 import { useRoute, useRouter } from 'vue-router';
 
-import AppTableStyled from '@/components/common/AppTableStyled.vue';
+import OrderDetailTable from '@/components/franchise/OrderDetailTable.vue';
+import PrintOrderPdfPreviewModal from '@/components/franchise/PrintOrderPdfPreviewModal.vue';
 import { useAppConfirmModal } from '@/hooks/useAppConfirmModal';
 import FCOrderApi from '@/utils/api/FCOrderApi';
 import { ORDER_STATUS } from '@/utils/constant';
@@ -86,6 +54,7 @@ const { showConfirm } = useAppConfirmModal();
 const toast = useToast();
 
 const orderDetail = ref(null);
+const showPrintPdf = ref(false);
 
 const isRequested = computed(() => {
   return orderDetail.value.orderStatus === ORDER_STATUS.REQUESTED;
@@ -97,16 +66,6 @@ const isCompleted = computed(() => {
 
 const fcOrderApi = new FCOrderApi();
 const { orderCode } = route.params;
-
-const totalPrice = computed(() => {
-  if (orderDetail.value && orderDetail.value.orderItemList) {
-    const totalPartSum = orderDetail.value.orderItemList.reduce((sum, item) => {
-      return sum + item.partSum + item.partSum * 0.1; // 부가세도 총합에 추가
-    }, 0);
-    return totalPartSum.toLocaleString();
-  }
-  return '0';
-});
 
 const clickExchange = () => {
   // TODO:: 교환 요청
@@ -135,6 +94,10 @@ const clickCancel = () => {
     onAccept: cancelOrder,
     danger: true,
   });
+};
+
+const clickPrintOrder = () => {
+  showPrintPdf.value = true;
 };
 
 onMounted(() => {
@@ -167,26 +130,6 @@ onMounted(() => {
     flex-direction: column;
     align-items: center;
     gap: 16px;
-
-    .approval-line-table {
-      align-self: flex-end;
-
-      th {
-        border: 1px solid black;
-        padding: 3px 0;
-        font-size: 13px;
-        width: 70px;
-      }
-      td {
-        border: 1px solid black;
-        font-size: 13px;
-        text-align: center;
-      }
-
-      & > tr:nth-child(2) {
-        height: 50px;
-      }
-    }
   }
 }
 </style>
